+++ pydantic_forms/validators/components/organisation_id.py
@@ -10,20 +10,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any
-from uuid import UUID
+from typing import Annotated
 
-from pydantic import GetCoreSchemaHandler, GetJsonSchemaHandler
-from pydantic.v1.validators import uuid_validator
-from pydantic_core import CoreSchema, core_schema
+from pydantic import Field
 
-
-class OrganisationId(UUID):
-    @classmethod
-    def __get_pydantic_json_schema__(cls, core_schema: CoreSchema, handler: GetJsonSchemaHandler) -> dict[str, Any]:
-        json_schema = handler.resolve_ref_schema(core_schema["schema"])
-        return json_schema | {"format": "organisationId"}
-
-    @classmethod
-    def __get_pydantic_core_schema__(cls, source_type: Any, handler: GetCoreSchemaHandler) -> CoreSchema:
-        return core_schema.no_info_after_validator_function(uuid_validator, handler(UUID))
+OrganisationId = Annotated[str, Field(json_schema_extra={"format": "organisationId"})]
